Stop bloating chain.ex + minor fixes
