+++ apps/explorer/lib/explorer/celo/events/contract_events/election/epoch_rewards_distributed_to_voters_event.ex
@@ -14,4 +14,13 @@ defmodule Explorer.Celo.ContractEvents.Election.EpochRewardsDistributedToVotersE
 
   event_param(:group, :address, :indexed)
   event_param(:value, {:uint, 256}, :unindexed)
+
+  def elected_groups_for_block(block_hash) do
+    events = query()
+             |> where([e], e.block_hash == ^block_hash)
+             |> Repo.all()
+             |> EventMap.celo_contract_event_to_concrete_event()
+
+    events |> Enum.map(& &1.group)
+  end
 end
